backend/requirements.txt
boto3
awscli
python-dotenv
amazon-textract-textractor[pdfium]
amazon-textract-textractor
pandas
pymupdf4llm
